Add files via upload
"This project involves data wrangling along with its analysis and is a part of my Doctoral thesis."

A ".lmd" file is used as an input file which is essentially a list mode data file wherein the data is written on event-by-event basis (time stamped), similar to a ".csv" file. 
The data file contains dynamical columns that are recorded for 105 events. 
The columns in the file are dynamically varying depending upon the non-zero values stored by thesisgorical variables "3000000000,3300000000.....,3333330000,5000000000,5500000000.....5555000000". In order to account for the detection efficiency of the measurement, data values are manipulated convoluting  with the uncertainties. 
The data is further analyzed with and without gating conditions to extract the desired results.